Move implementation of SSE4.1 dpps/dppd to helper function
